py/requirements.txt
azure-storage-blob
flask
flask-cors
python-dotenv
reportlab
requests
google
Pillow